refactor: Move BackupDone in own component
