+++ src/photos/ducks/backup/components/BackupInfo.jsx
@@ -1,25 +1,12 @@
 import React from 'react'
-import cx from 'classnames'
-
-import Alert from 'cozy-ui/transpiled/react/Alert'
-import AlertTitle from 'cozy-ui/transpiled/react/AlertTitle'
-import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
 
 import AnalysisInProgress from 'photos/ducks/backup/components/BackupInfo/AnalysisInProgress'
 import BackupReady from 'photos/ducks/backup/components/BackupInfo/BackupReady'
+import BackupDone from 'photos/ducks/backup/components/BackupInfo/BackupDone'
 
-import styles from '../../../styles/backup.styl'
 import { useBackupData } from '../hooks/useBackupData'
 
-const formatLastBackupMessage = message => {
-  if (!message) return ''
-
-  return message.charAt(0).toUpperCase() + message.slice(1) + '. '
-}
-
 const BackupInfo = () => {
-  const { t } = useI18n()
-
   const { backupInfo } = useBackupData()
 
   if (!backupInfo) return null
@@ -34,29 +21,7 @@ const BackupInfo = () => {
   } else if (status === 'ready') {
     return <BackupReady mediasToBackupCount={mediasToBackupCount} />
   } else if (status === 'done') {
-    if (lastBackup.status === 'success') {
-      return (
-        <div className={cx('u-mt-1-half', styles['pho-backup-info-wrapper'])}>
-          <Alert severity="success">
-            <AlertTitle>{t('Backup.info.successTitle')}</AlertTitle>
-            {t('Backup.info.successDescription')}
-          </Alert>
-        </div>
-      )
-    } else {
-      return (
-        <div className={cx('u-mt-1-half', styles['pho-backup-info-wrapper'])}>
-          <Alert severity="error">
-            <AlertTitle>{t('Backup.info.errorTitle')}</AlertTitle>
-            {formatLastBackupMessage(lastBackup.message)}
-            {t('Backup.info.errorDescription', {
-              smart_count: lastBackup.totalMediasToBackupCount,
-              backedUpMediaCount: lastBackup.backedUpMediaCount
-            })}
-          </Alert>
-        </div>
-      )
-    }
+    return <BackupDone lastBackup={lastBackup} />
   }
 
   return null
